+++ tests/TestCase/TestSuite/TestTraitTest.php
@@ -189,7 +189,7 @@ class TestTraitTest extends TestCase
      */
     public function testAssertFileMime()
     {
-        $this->assertFileMime(create_tmp_file('string'), 'text/plain');
+        $this->assertFileMime('text/plain', create_tmp_file('string'));
     }
 
     /**