+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 1.x branch
 ## 1.2 branch
 ### 1.2.0
+* arguments of the `assertFileMime()` assert method have been reversed
+    (`$expectedMime, $filename, $message = ''`);
 * removed deprecated `ends_with()` and `starts_with()` global functions, use
     instead `string_ends_with()` and `string_starts_with()`;
 * removed deprecated `first_key()`, `first_value()`, `first_value_recursive()`,